refactor: split git_machete/utils.py into a package with submodules
The 580-line utils.py covered ten loosely related concerns: TTY
detection, ANSI codes, markup formatting, path helpers, generic
collections, file system, debug logging, command execution, and
exceptions/enums. Split it into a `git_machete/utils/` package with one
submodule per concern, and keep `git_machete/utils/__init__.py` as a
re-export shim so every existing `from git_machete.utils import X` and
`utils.X` call site keeps working unchanged.

Layering (each module imports only from those above it in the list):

  _subproc           : bare subprocess wrappers (no logging)
  paths              : POSIX-style path helpers
  collections_utils  : iterable/sequence helpers
  terminal           : TTY detection + ANSI code classes
  markup             : markup language and styled output
  debug_log          : debug() with arg redaction
  fs                 : file-system helpers
  cmd                : run_cmd / popen_cmd (with verbose/debug logging)
  exceptions         : MacheteException et al., ExitCode, ParsableEnum

The mutable runtime flags (`debug_mode`, `verbose_mode`,
`use_ansi_escapes_in_*`, `current_directory_confirmed_to_exist`,
`measure_command_time`) and the token-redaction constants stay on the
package object so external `cli.py` writes (`utils.verbose_mode = True`)
keep propagating to every reader, including submodules. Submodules read
those flags via a function-local `from git_machete import utils as
_utils; _utils.X` - lazy on purpose so pylint's `cyclic-import` checker
doesn't see a back-edge from the submodule to the package, and so that
`mock.patch('git_machete.utils.X', ...)` in tests is honoured (the call
sites resolve `_utils.X` at call time, after the patch is in effect).

Lint/test infra:
- `__all__` in `__init__.py` documents the public surface and prevents
  `autoflake --remove-all-unused-imports` (run in the `isort` testenv)
  from stripping the re-export imports.
- `cyclic-import-check` adds `--ignore-paths=git_machete/utils.*` to
  silence pylint's false positive: it treats every `from .X import Y`
  re-export in `__init__.py` as a back-edge from the package to its
  submodule, then pairs that with each `from .Y import Z` between
  submodules to "form" a cycle - even though no submodule imports the
  parent package at module level. The runtime parent-package imports
  are deliberately function-local for exactly this reason.

All 419 tests pass and all six lint envs (flake8-check, isort-check,
mypy, vulture-check, cyclic-import-check, pep8) are clean.
