+++ tox.ini
@@ -178,8 +178,14 @@ commands = vulture --ignore-names check_hostname,verify_mode,pytest_*,_check_val
 description = "Detect circular imports"
 deps =
   -r{[requirements]dir}/cyclic-import-check.txt
+# `git_machete/utils/` is excluded because pylint's `cyclic-import` checker emits a false positive for any package
+# whose `__init__.py` re-exports symbols from its own submodules: it treats the `from .X import ...` re-export as a
+# back-edge from the package to its submodule, which it then pairs with each `from .Y import ...` between submodules
+# to "form" a cycle - even though no submodule imports the parent at module level. The actual runtime parent-package
+# imports inside `git_machete.utils.{cmd,markup,debug_log,exceptions}` are deliberately function-local for exactly this
+# reason (and to keep `mock.patch('git_machete.utils.X', ...)` working in tests).
 commands =
-  pylint --disable=all --enable=cyclic-import git_machete/
+  pylint --disable=all --enable=cyclic-import --ignore-paths=git_machete/utils.* git_machete/
 
 
 ## DOCS
